Save content to "./tmp" or "/tmp"
This commit updates the `SaveContent` function to first check if the code is running in a Lambda function. If it is, the content is saved to the "/tmp" directory. If it is not running in Lambda, it checks if the "./tmp" directory exists. If it does not exist, it creates the directory before saving the content.